Simplified layer comps rename
